requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
langchain-community
faiss-cpu
beautifulsoup4
requests
lxml
pydantic-settings
yfinance
